Strict casts
perfstats.c: r->mu is the mean, so doing the calculation with
    floating-point makes sense.

hazenquantile.c: the goal of this assert() is to ensure that we don't
    end up with
        i == N-1 and r > 0
    In other words, if N*x is large enough, we need to end up with
        i == N-1, r == 0.0
    to avoid the client code accessing S_N (which does not exist).

    As such, we need to do this math with doubles, because if we did
    (size_t)*r, then the assert() would be happy with r == 0.5, but that
    would break client code.

+++ perf/perfstats.c
@@ -263,7 +263,7 @@ perfstats_add(struct perfstats * P, uint32_t reqtype, double t)
 	/* Add to the statistics for this request type. */
 	r = *reqstatslist_get(P->stats, i);
 	r->N++;
-	r->mu += (t - r->mu) / r->N;
+	r->mu += (t - r->mu) / (double)r->N;
 	if (onlinequantile_add(r->p50, t))
 		goto err0;
 	if (onlinequantile_add(r->p99, t))
